+++ src/Yaml/LibYaml.php
@@ -13,13 +13,15 @@ class LibYaml implements ParserInterface
      */
     public static function isAvailable()
     {
-        return extension_loaded('yaml') && function_exists('yaml_parse_file') && function_exists('yaml_parse');
+        return extension_loaded('yaml') && function_exists('yaml_parse');
     }
 
     /**
      * @param string $path
      *
-     * @return null|array
+     * @throws ParseException
+     *
+     * @return array
      */
     public function parseFile($path)
     {
@@ -27,19 +29,57 @@ class LibYaml implements ParserInterface
     }
 
     /**
+     * @param string $path
+     *
+     * @return null|array
+     */
+    public function tryParseFile($path)
+    {
+        return Yaml::fileDelegate($path, array($this, 'tryParseBuffer'));
+    }
+
+    /**
+     * @param string $buffer
+     *
+     * @throws ParseException
+     *
+     * @return array
+     */
+    public function parseBuffer($buffer)
+    {
+        if (!function_exists('yaml_parse')) {
+            // @codeCoverageIgnoreStart
+            throw new \BadMethodCallException('LibYaml based parsing n/a, is the PHP extension loaded?');
+            // @codeCoverageIgnoreEnd
+        }
+
+        $error = ErrorCatcher::create();
+        $result = yaml_parse($buffer, 0);
+        $result = $error->end() ? null : $result;
+        if (null === $result && null !== $message = $error->getLastErrorMessage()) {
+            throw new ParseException($message);
+        }
+        if (!is_array($result)) {
+            throw new ParseException('LibYaml invalid YAML parsing');
+        }
+
+        # ext-yaml parser does aliases, remove any potential ones
+        return json_decode(json_encode($result), true);
+    }
+
+    /**
      * @param string $buffer
      *
      * @return null|array
      */
-    public function parseBuffer($buffer)
+    public function tryParseBuffer($buffer)
     {
-        $error = ErrorCatcher::create();
-        $result = yaml_parse($buffer, 0);
-        $result = $error->end() ? null : $result;
+        try {
+            $result = $this->parseBuffer($buffer);
+        } catch (ParseException $ex) {
+            return null;
+        }
 
-        return !is_array($result)
-            ? null
-            # ext-yaml parser does aliases, remove any potential ones
-            : json_decode(json_encode($result), true);
+        return $result;
     }
 }