+++ src/Yaml/ParseException.php
@@ -1,0 +1,9 @@
+<?php
+
+/* this file is part of pipelines */
+
+namespace Ktomk\Pipelines\Yaml;
+
+class ParseException extends \RuntimeException
+{
+}