+++ src/Yaml/ParserInterface.php
@@ -14,14 +14,32 @@ interface ParserInterface
     /**
      * @param string $path
      *
-     * @return null|array
+     * @throws ParseException
+     *
+     * @return array
      */
     public function parseFile($path);
+
+    /**
+     * @param string $path
+     *
+     * @return null|array
+     */
+    public function tryParseFile($path);
+
+    /**
+     * @param string $buffer
+     *
+     * @throws ParseException
+     *
+     * @return array
+     */
+    public function parseBuffer($buffer);
 
     /**
      * @param string $buffer
      *
      * @return null|array
      */
-    public function parseBuffer($buffer);
+    public function tryParseBuffer($buffer);
 }