+++ CHANGELOG.md
@@ -9,11 +9,14 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 
 ## [unreleased]
 ### Change
+- Enrich YAML parse error information (#14) (thanks [Adrian Cederberg])
 - html-docs: Upgrade mkdocs-material to 8.1.3
 ### Fix
 - Fix composer version in phar build info for PHP 5.3 (and up sometimes),
   since [0.0.51](#0051---2020-12-09)
 - Fix portable path check, since [0.0.42](#0042---2020-06-25)
+
+[Adrian Cederberg]: https://github.com/acederberg
 
 ## [0.0.62] - 2021-12-19
 ### Add